+++ library/systemd/examples/comparison_of_widgets.rb
@@ -2,9 +2,8 @@ require "yast"
 
 require "yast2/service_widget"
 require "yast2/service_configuration"
+require "yast2/systemd_service"
 require "ui/service_status"
-
-Yast.import "SystemdService"
 
 def service
   @service ||= Yast::SystemdService.find!("cups.service")
